requirements.txt
Django==1.6.1
South==0.8.4
argparse==1.2.1
django-classy-tags==0.4
django-cms==2.4.3
django-mptt==0.6.0
django-sekizai==0.7
djangocms-admin-style==0.1.2
html5lib==1.0b1
six==1.5.2
wsgiref==0.1.2
